+++ Unity/RTS-Game/Assets/Daniel/PauseMenuScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2300e862877969443993451c4a1a29cf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
