+++ Unity/RTS-Game/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.4.11f1
-m_EditorVersionWithRevision: 2019.4.11f1 (2d9804dddde7)
+m_EditorVersion: 2019.3.0f5
+m_EditorVersionWithRevision: 2019.3.0f5 (44796c9d3c2c)
